requirements.txt
flask
flask-socketio
flask-cors
opencv-python
insightface
onnxruntime
numpy
Pillow
eventlet
